backend/requirements.txt
# requirements.txt - StudyGenie Backend
# Testé et validé - 2026-01-28

# ============================================
# CORE FRAMEWORK
# ============================================
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-dotenv==1.0.0
pydantic==2.10.3
pydantic-settings==2.7.0

# ============================================
# DATABASE
# ============================================
SQLAlchemy==2.0.45

# ============================================
# AUTHENTICATION & SECURITY
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
PyJWT==2.10.1
email-validator==2.3.0

# ============================================
# AI & ML
# ============================================
anthropic==0.40.0
openai==1.58.1
voyageai==0.2.3
faiss-cpu==1.13.2

# ============================================
# DOCUMENT PROCESSING
# ============================================
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
reportlab==4.0.9
Pillow==12.1.0
pytesseract==0.3.13

# ============================================
# UTILS
# ============================================
requests==2.31.0
typing-extensions==4.15.0

# ============================================
# PRODUCTION
# ============================================
gunicorn==24.0.0

# ============================================
# PAYMENTS
# ============================================
stripe==14.1.0

# ============================================
# NOTES
# ============================================
# Installation : pip install -r requirements.txt
# Python version : 3.10+
# OS testé : Windows 11
# Tesseract requis pour OCR : https://github.com/tesseract-ocr/tesseract
